Rename module 'util' to 'common' to better reflect its usage
