+++ src/sdl2/rect.rs
@@ -2,7 +2,7 @@ use sys::rect as ll;
 use std::mem;
 use std::ptr;
 use std::ops::{BitAnd, BitOr};
-use util::validate_int;
+use common::validate_int;
 use get_error;
 
 /// Immutable point type, consisting of x and y.
